fix(events): address Claude Code Review findings
Three findings from the PR 283 auto-review:

1. Shared repo helpers moved to the right file. uuidPtrToPgUUID,
   pgTextToStrPtr, and pgTimestamptzToTimePtr belong in conversions.go
   with the other helpers, not embedded in event_shadow_observation.go
   (per core.md gotcha "Redefining helper functions in new repository
   files").

2. TestIntegration_PublishTxEnqueuesInteractionRecorderJob now uses
   before/after counts to assert exactly one job was enqueued by THIS
   publish, not a global >=1 that would pass from leftover rows.

3. Remove dead `_ = json.RawMessage(raw)` line in buildManualEnvelope.
   json.RawMessage is `type RawMessage []byte` — the conversion was a
   no-op that did no validation. events.Marshal already returns
   guaranteed-valid JSON via json.Marshal, so no replacement needed.
