+++ backend/internal/repository/conversions.go
@@ -31,3 +31,31 @@ func timeToPgTimestamptz(t *time.Time) pgtype.Timestamptz {
 	}
 	return pgtype.Timestamptz{Time: *t, Valid: true}
 }
+
+// uuidPtrToPgUUID converts a *uuid.UUID to pgtype.UUID, treating nil as
+// invalid (NULL). Used for nullable FK / ref columns.
+func uuidPtrToPgUUID(id *uuid.UUID) pgtype.UUID {
+	if id == nil {
+		return pgtype.UUID{Valid: false}
+	}
+	return pgtype.UUID{Bytes: *id, Valid: true}
+}
+
+// pgTextToStrPtr converts a pgtype.Text to *string; invalid/NULL → nil.
+func pgTextToStrPtr(t pgtype.Text) *string {
+	if !t.Valid {
+		return nil
+	}
+	s := t.String
+	return &s
+}
+
+// pgTimestamptzToTimePtr converts a pgtype.Timestamptz to *time.Time;
+// invalid/NULL → nil. Returned time is normalized to UTC.
+func pgTimestamptzToTimePtr(t pgtype.Timestamptz) *time.Time {
+	if !t.Valid {
+		return nil
+	}
+	u := t.Time.UTC()
+	return &u
+}
